--- app/helpers/profile_helper.rb
@@ -1,2 +1,0 @@
-module ProfileHelper
-end
