+++ config/routes.rb
@@ -2,7 +2,7 @@ Rails.application.routes.draw do
   root 'home#index'
 
   get "up" => "rails/health#show", as: :rails_health_check
-  get 'profile', to: 'profile#profile'
+  get 'profiles', to: 'profiles#profile'  # Update here
 
   resources :profiles, only: [:index, :show, :create]
 end
