+++ test/controllers/home_controller_test.rb
@@ -1,7 +1,7 @@
 require "test_helper"
 
 class HomeControllerTest < ActionDispatch::IntegrationTest
-  test "should get profile" do
+  test "should get profiles" do
     get home_index_url
     assert_response :success
   end
